chore(carousel): move carousel out of legacy into components
